refactor: :construction: update go version. 1. update go version. 2. correction code directory
