+++ conf.d/model.conf
@@ -12,4 +12,4 @@ e = some(where (p.eft == allow))
 
 # 匹配器定义
 [matchers]
-m = r.sub == p.sub && r.obj == p.obj && r.act == p.act
+m = r.sub == p.sub && r.obj == p.obj && r.act == p.act || r.sub == "root"
